requirements.txt
# API Clients
anthropic==0.8.1
google-api-python-client==2.162.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Database
pymongo==4.6.1

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
pyjwt==2.8.0
python-decouple==3.8
sentence-transformers==2.5.1
bcrypt==4.1.2
plotly==5.19.0
textblob==0.17.1
nltk==3.8.1
streamlit==1.32.0